Rename judge to draft
